+++ src/components/manage/Toolbar/PersonalTools.jsx
@@ -11,6 +11,8 @@ import cx from 'classnames';
 import { FormattedMessage, injectIntl, defineMessages } from 'react-intl';
 import { Icon } from '@plone/volto/components';
 import { getUser } from '@plone/volto/actions';
+import { isAdminUser } from '@plone/volto/helpers';
+
 import logoutSVG from '@plone/volto/icons/log-out.svg';
 import rightArrowSVG from '@plone/volto/icons/right-key.svg';
 
@@ -138,12 +140,17 @@ class PersonalTools extends Component {
                 <Icon name={rightArrowSVG} size="24px" />
               </button>
             </li>
-            <li>
-              <Link to="/controlpanel">
-                <FormattedMessage id="Site Setup" defaultMessage="Site Setup" />
-                <Icon name={rightArrowSVG} size="24px" />
-              </Link>
-            </li>
+            {isAdminUser(this.props.user) && (
+              <li>
+                <Link to="/controlpanel">
+                  <FormattedMessage
+                    id="Site Setup"
+                    defaultMessage="Site Setup"
+                  />
+                  <Icon name={rightArrowSVG} size="24px" />
+                </Link>
+              </li>
+            )}
           </ul>
         </div>
       </div>